backend/requirements.txt
# RiskGuard v3 — Python dependencies
# pip install -r requirements.txt

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── ML / Data ─────────────────────────────────────────────────────────────────
numpy>=1.26.0
scipy>=1.13.0          # audio loading (wavfile) + signal processing
Pillow>=10.3.0         # image processing

# ── Optional but recommended ──────────────────────────────────────────────────
# Install for faster/better results:
opencv-python-headless>=4.9.0    # DCT (image) + optical flow (video)
# librosa>=0.10.1                # Better audio analysis (needs ffmpeg)